web-app/requirements.txt
Pillow
azure-storage-blob==12.6.0
flask-appbuilder
